Added startup systemd, change compiling setting and more
-Added compiling optimization
-Fix warning message compiling with -Wall
-Added support AMBE board TEAM6160
-Added file startup systemd
-Updated file README.md

+++ ambed/ccallsign.cpp
@@ -93,7 +93,7 @@ void CCallsign::SetCallsign(const uint8 *buffer, int len)
     // set callsign
     ::memset(m_Callsign, ' ', sizeof(m_Callsign));
     ::memcpy(m_Callsign, buffer, MIN(len, sizeof(m_Callsign)));
-    for ( int i = 0; i < sizeof(m_Callsign); i++ )
+    for ( size_t i = 0; i < sizeof(m_Callsign); i++ )
     {
         if ( m_Callsign[i] == 0 )
         {
@@ -108,7 +108,7 @@ void CCallsign::SetCallsign(const uint8 *buffer, int len)
 
 void CCallsign::PatchCallsign(int off, const uint8 *patch, int len)
 {
-    if ( off < sizeof(m_Callsign) )
+    if ( (size_t)off < sizeof(m_Callsign) )
     {
         ::memcpy(m_Callsign, patch, MIN(len, sizeof(m_Callsign) - off));
     }
@@ -125,7 +125,7 @@ void CCallsign::GetCallsign(uint8 *buffer) const
 
 void CCallsign::GetCallsignString(char *sz) const
 {
-    int i;
+    size_t i;
     for ( i = 0; (i < sizeof(m_Callsign)) && (m_Callsign[i] != ' '); i++ )
     {
         sz[i] = m_Callsign[i];
@@ -146,7 +146,7 @@ bool CCallsign::HasSameCallsignWithWildcard(const CCallsign &callsign) const
     bool same = true;
     bool done = false;
     
-    for ( int i = 0; (i < sizeof(m_Callsign)) && same && !done; i++ )
+    for ( size_t i = 0; (i < sizeof(m_Callsign)) && same && !done; i++ )
     {
         if ( !(done = ((m_Callsign[i] == '*') || (callsign[i] == '*'))) )
         {
