+++ src/creflector.cpp
@@ -568,7 +568,7 @@ CPacketStream *CReflector::GetStream(char module)
 bool CReflector::IsStreamOpen(const CDvHeaderPacket *DvHeader)
 {
     bool open = false;
-    for ( int i = 0; (i < m_Streams.size()) && !open; i++  )
+    for ( size_t i = 0; (i < m_Streams.size()) && !open; i++  )
     {
         open =  ( (m_Streams[i].GetStreamId() == DvHeader->GetStreamId()) &&
                   (m_Streams[i].IsOpen()));
@@ -579,7 +579,7 @@ bool CReflector::IsStreamOpen(const CDvHeaderPacket *DvHeader)
 char CReflector::GetStreamModule(CPacketStream *stream)
 {
     char module = ' ';
-    for ( int i = 0; (i < m_Streams.size()) && (module == ' '); i++ )
+    for ( size_t i = 0; (i < m_Streams.size()) && (module == ' '); i++ )
     {
         if ( &(m_Streams[i]) == stream )
         {
