+++ src/ctranscoder.cpp
@@ -65,7 +65,7 @@ CTranscoder::~CTranscoder()
     // close all streams
     m_Mutex.lock();
     {
-        for ( int i = 0; i < m_Streams.size(); i++ )
+        for ( size_t i = 0; i < m_Streams.size(); i++ )
         {
             delete m_Streams[i];
         }
@@ -116,11 +116,11 @@ void CTranscoder::Close(void)
 {
     // close socket
     m_Socket.Close();
-    
+
     // close all streams
     m_Mutex.lock();
     {
-        for ( int i = 0; i < m_Streams.size(); i++ )
+        for ( size_t i = 0; i < m_Streams.size(); i++ )
         {
             delete m_Streams[i];
         }
@@ -128,7 +128,7 @@ void CTranscoder::Close(void)
         
     }
     m_Mutex.unlock();
-    
+ 
     // kill threads
     m_bStopThread = true;
     if ( m_pThread != NULL )
@@ -254,8 +254,7 @@ CCodecStream *CTranscoder::GetStream(CPacketStream *PacketStream, uint8 uiCodecI
             else
             {
                 std::cout << "ambed openstream timeout" << std::endl;
-            }
-            
+            }           
         }
     }
     return stream;
@@ -271,7 +270,7 @@ void CTranscoder::ReleaseStream(CCodecStream *stream)
         bool found = false;
         Lock();
         {
-            for ( int i = 0; (i < m_Streams.size()) && !found; i++ )
+            for ( size_t i = 0; (i < m_Streams.size()) && !found; i++ )
             {
                 // compare object pointers
                 if ( (m_Streams[i]) ==  stream )
